Big refactors. Drops getting the tenant/ARA from EstadoCollection
Now, the method that must be used in front to get which tenant is running the app must be:

`window.SIRHA.getARA` or `window.SIRHA.is_single_mode`

Old var `window.SIRHA.ARA` is rename to `window.SIRHA._ARA` to force the refactoring of
the existing code. The reason of this change, is that is easier search for `getARA`
occurrences in the code base that for `ARA`.

Also, some previous infortunate use of `estados.fetch` just to get the ara is dropped