Moves column `c_estimado` from child tables to `utentes.actividade` parent
In some states of the renovation and granting process the value of `c_estimado`
is need to validate that the exp have all the needed fields.

To avoid a join between all the child tables in the list views, c_estimado must
be in the actividade table.

Also, as now we are not sending the full Exploracao entity in all lists-views
specific field updates are need in some cases

refs #1516
refs #2096